Fix a few XML Bugs Increase precision for FPM

+++ cgf-converter/Structs.cs
@@ -55,10 +55,10 @@ namespace CgfConverter
     } // String32 Name, int Start, int End - complete
     public struct Vector3
     {
-        public float x;
-        public float y;
-        public float z;
-        public float w; // Currently Unused
+        public Double x;
+        public Double y;
+        public Double z;
+        public Double w; // Currently Unused
         public void ReadVector3(BinaryReader b)
         {
             this.x = b.ReadSingle();
@@ -93,10 +93,10 @@ namespace CgfConverter
     }  // Vector in 3D space {x,y,z}
     public struct Vector4
     {
-        public float x;
-        public float y;
-        public float z;
-        public float w;
+        public Double x;
+        public Double y;
+        public Double z;
+        public Double w;
 
         public Vector3 ToVector3()
         {
@@ -122,15 +122,15 @@ namespace CgfConverter
     }
     public struct Matrix33    // a 3x3 transformation matrix
     {
-        public float m11;
-        public float m12;
-        public float m13;
-        public float m21;
-        public float m22;
-        public float m23;
-        public float m31;
-        public float m32;
-        public float m33;
+        public Double m11;
+        public Double m12;
+        public Double m13;
+        public Double m21;
+        public Double m22;
+        public Double m23;
+        public Double m31;
+        public Double m32;
+        public Double m33;
 
         public void ReadMatrix33(BinaryReader b)
         {
@@ -172,7 +172,7 @@ namespace CgfConverter
             mat.m33 = m33;
             return mat;
         }
-        public float Get_Determinant()
+        public Double Get_Determinant()
         {
             return (m11 * m22 * m33
                   + m12 * m23 * m31
@@ -240,9 +240,9 @@ namespace CgfConverter
             // Get the scale, assuming is_scale_rotation is true
             Matrix33 mat = this.Mult(this.Get_Transpose());
             Vector3 scale = new Vector3();
-            scale.x = (float)System.Math.Pow(mat.m11, 0.5);
-            scale.y = (float)System.Math.Pow(mat.m22, 0.5);
-            scale.z = (float)System.Math.Pow(mat.m33, 0.5);
+            scale.x = (Double)System.Math.Pow(mat.m11, 0.5);
+            scale.y = (Double)System.Math.Pow(mat.m22, 0.5);
+            scale.z = (Double)System.Math.Pow(mat.m33, 0.5);
             if (this.Get_Determinant() < 0)
             {
                 scale.x = 0 - scale.x;
@@ -283,31 +283,31 @@ namespace CgfConverter
     }
     public struct Matrix44    // a 4x4 transformation matrix.  first value is row, second is column
     {
-        public float m11;
-        public float m12;
-        public float m13;
-        public float m14;
-        public float m21;
-        public float m22;
-        public float m23;
-        public float m24;
-        public float m31;
-        public float m32;
-        public float m33;
-        public float m34;
-        public float m41;
-        public float m42;
-        public float m43;
-        public float m44;
+        public Double m11;
+        public Double m12;
+        public Double m13;
+        public Double m14;
+        public Double m21;
+        public Double m22;
+        public Double m23;
+        public Double m24;
+        public Double m31;
+        public Double m32;
+        public Double m33;
+        public Double m34;
+        public Double m41;
+        public Double m42;
+        public Double m43;
+        public Double m44;
 
         public Vector4 Mult4x1(Vector4 vector)
         {
             // Pass the matrix a Vector4 (4x1) vector to get the transform of the vector
             Vector4 result = new Vector4();
-            //result.x = (m11 * vector.x) + (m12 * vector.y) + (m13 * vector.z) + m14/100;
-            //result.y = (m21 * vector.x) + (m22 * vector.y) + (m23 * vector.z) + m24/100;
-            //result.z = (m31 * vector.x) + (m32 * vector.y) + (m33 * vector.z) + m34/100;
-            //result.w = (m41 * vector.x) + (m42 * vector.y) + (m43 * vector.z) + m44/100;
+            // result.x = (m11 * vector.x) + (m12 * vector.y) + (m13 * vector.z) + m14 / 100;
+            // result.y = (m21 * vector.x) + (m22 * vector.y) + (m23 * vector.z) + m24 / 100;
+            // result.z = (m31 * vector.x) + (m32 * vector.y) + (m33 * vector.z) + m34 / 100;
+            // result.w = (m41 * vector.x) + (m42 * vector.y) + (m43 * vector.z) + m44 / 100;
             result.x = (m11 * vector.x) + (m21 * vector.y) + (m31 * vector.z) + m41 / 100;
             result.y = (m12 * vector.x) + (m22 * vector.y) + (m32 * vector.z) + m42 / 100;
             result.z = (m13 * vector.x) + (m23 * vector.y) + (m33 * vector.z) + m43 / 100;
@@ -351,10 +351,10 @@ namespace CgfConverter
 
     public struct Quat        // A quaternion (x,y,z,w)
     {
-        public float x;
-        public float y;
-        public float z;
-        public float w;
+        public Double x;
+        public Double y;
+        public Double z;
+        public Double w;
     }
     public struct Vertex      // position p(Vector3) and normal n(Vector3)
     {
@@ -376,7 +376,7 @@ namespace CgfConverter
         public uint FirstVertex;
         public uint NumVertices;
         public uint MatID;
-        public float Radius;
+        public Double Radius;
         public Vector3 Center;
     }  // Contains data about the parts of a mesh, such as vertices, radius and center.
     public struct Key
@@ -386,12 +386,12 @@ namespace CgfConverter
         public Vector3 RelPos; // relative position
         public Quat RelQuat; //Relative Quaternion if ARG==1?
         public Vector3 Unknown1; // If ARG==6 or 10?
-        public float[] Unknown2; // If ARG==9?  array length = 2
+        public Double[] Unknown2; // If ARG==9?  array length = 2
     }
     public struct UV
     {
-        public float U;
-        public float V;
+        public Double U;
+        public Double V;
     }
     public struct UVFace
     {
@@ -420,9 +420,9 @@ namespace CgfConverter
     }
     public struct FRGB
     {
-        public float r; // float Red
-        public float g; // float green
-        public float b; // float blue
+        public Double r; // Double Red
+        public Double g; // Double green
+        public Double b; // Double blue
     }
     public struct Tangent
     {
@@ -434,11 +434,11 @@ namespace CgfConverter
     }
     public struct WORLDTOBONE
     {
-        public float[,] worldToBone;   //  4x3 structure
+        public Double[,] worldToBone;   //  4x3 structure
         
         public void GetWorldToBone(BinaryReader b)
         {
-            worldToBone = new float[4,3];
+            worldToBone = new Double[4,3];
             //Console.WriteLine("GetWorldToBone {0:X}", b.BaseStream.Position);
             for (int i = 0; i<4; i++) 
             {
@@ -464,12 +464,12 @@ namespace CgfConverter
     }
     public struct BONETOWORLD
     {
-        public float[,] boneToWorld;   //  4x3 structure
+        public Double[,] boneToWorld;   //  4x3 structure
 
         public void GetBoneToWorld(BinaryReader b)
         {
             //BONETOWORLD tempB2W = new BONETOWORLD();
-            boneToWorld = new float[4, 3];
+            boneToWorld = new Double[4, 3];
             //Console.WriteLine("GetBoneToWorld");
             for (int i = 0; i < 4; i++)
             {
@@ -529,7 +529,7 @@ namespace CgfConverter
     {
         public UInt32 controllerID;
         public PhysicsGeometry[] physicsGeometry; // 2 of these.
-        public float mass;                  // 0xD8 ?
+        public Double mass;                  // 0xD8 ?
         public WORLDTOBONE worldToBone;     // 4x3 matrix
         public BONETOWORLD boneToWorld;     // 4x3 matrix
         public String boneName;             // String256 in old terms; convert to a real null terminated string.
